Refactor code to have logic in card container and a lazy-load on card itself
